Remove some obsolete lanes +  reorder for consistency
Now that the build and the upload are done by separate jobs on CI (both for Prototype Builds in `pipeline.yml` and for Beta/Final builds in `release-builds.yml`), we don't need the previous lanes that did both in one go.

IIRC back when we split the lanes in 2 jobs it was argued that it could be useful to keep the old lanes around in case one would need to run them locally.
But I'd argue that this makes us keep some dead/legacy code around that won't be maintained, and is likely to go stale, and clutters the Fastfile and makes it less easy to navigate and reason about, so I think it's clearer if we remove this unused code after all.